# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fredora Multiconcept API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ServiceItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  divisionSlug: string;
  sortOrder: number;
  createdAt: string;
}

export interface Division {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  tagline?: string | null;
  description: string;
  bannerColor: string;
  comingSoon: boolean;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
  services: ServiceItem[];
  createdAt: string;
  updatedAt: string;
}

export interface UpdateDivisionBody {
  name?: string;
  /** @nullable */
  tagline?: string | null;
  description?: string;
  bannerColor?: string;
  comingSoon?: boolean;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Homepage {
  id: number;
  heroTitle: string;
  heroSubtitle: string;
  motto: string;
  missionStatement: string;
  visionStatement: string;
  coreValues: string[];
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  googleAnalyticsId?: string | null;
  updatedAt: string;
}

export interface UpdateHomepageBody {
  heroTitle?: string;
  heroSubtitle?: string;
  motto?: string;
  missionStatement?: string;
  visionStatement?: string;
  coreValues?: string[];
  /** @nullable */
  heroImageUrl?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  googleAnalyticsId?: string | null;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface GalleryItem {
  id: number;
  /** @nullable */
  divisionSlug?: string | null;
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CreateGalleryItemBody {
  /** @nullable */
  divisionSlug?: string | null;
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  sortOrder?: number;
}

export interface Product {
  id: number;
  divisionSlug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  price?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CreateProductBody {
  divisionSlug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  price?: string | null;
  sortOrder?: number;
}

export interface Testimonial {
  id: number;
  authorName: string;
  /** @nullable */
  company?: string | null;
  content: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  divisionSlug?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CreateTestimonialBody {
  authorName: string;
  /** @nullable */
  company?: string | null;
  content: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  divisionSlug?: string | null;
  sortOrder?: number;
}

export interface NewsletterSubscribeBody {
  email: string;
}

export interface NewsletterSubscriber {
  id: number;
  email: string;
  createdAt: string;
}

export interface Message {
  id: number;
  name: string;
  email: string;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface CreateMessageBody {
  name: string;
  email: string;
  message: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  username: string;
}

export interface AdminSession {
  loggedIn: boolean;
  /** @nullable */
  username?: string | null;
}

export interface AdminStats {
  totalDivisions: number;
  totalMessages: number;
  unreadMessages: number;
  totalServices: number;
}

export type ListGalleryItemsParams = {
  divisionSlug?: string;
};

export type ListProductsParams = {
  divisionSlug?: string;
};