# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fredora Multiconcept API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: divisions
    description: Business divisions
  - name: homepage
    description: Homepage content
  - name: messages
    description: Contact messages
  - name: admin
    description: Admin operations
  - name: Storage
    description: Object storage upload and serving endpoints.
  - name: gallery
    description: Gallery / photo items
  - name: products
    description: Product catalog
  - name: testimonials
    description: Customer testimonials
  - name: newsletter
    description: Newsletter subscriptions

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /divisions:
    get:
      operationId: listDivisions
      tags: [divisions]
      summary: List all divisions
      responses:
        "200":
          description: List of divisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Division"

  /divisions/{slug}:
    get:
      operationId: getDivision
      tags: [divisions]
      summary: Get a specific division by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Division detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Division"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      operationId: updateDivision
      tags: [divisions]
      summary: Update a division (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDivisionBody"
      responses:
        "200":
          description: Updated division
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Division"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /homepage:
    get:
      operationId: getHomepage
      tags: [homepage]
      summary: Get homepage content
      responses:
        "200":
          description: Homepage content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Homepage"

    patch:
      operationId: updateHomepage
      tags: [homepage]
      summary: Update homepage content (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHomepageBody"
      responses:
        "200":
          description: Updated homepage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Homepage"

  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List contact messages (admin)
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: createMessage
      tags: [messages]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageBody"
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a contact message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Admin session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /gallery:
    get:
      operationId: listGalleryItems
      tags: [gallery]
      summary: List gallery items
      parameters:
        - name: divisionSlug
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Add a gallery item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryItemBody"
      responses:
        "201":
          description: Created gallery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"

  /gallery/{id}:
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete a gallery item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: divisionSlug
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Add a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Add a testimonial (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestimonialBody"
      responses:
        "201":
          description: Created testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"

  /testimonials/{id}:
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete a testimonial (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscribeBody"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "409":
          description: Already subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /newsletter/subscribers:
    get:
      operationId: listNewsletterSubscribers
      tags: [newsletter]
      summary: List newsletter subscribers (admin)
      responses:
        "200":
          description: List of subscribers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsletterSubscriber"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    ServiceItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        divisionSlug:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - divisionSlug
        - sortOrder
        - createdAt

    Division:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        tagline:
          type: ["string", "null"]
        description:
          type: string
        bannerColor:
          type: string
        comingSoon:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - name
        - description
        - bannerColor
        - comingSoon
        - sortOrder
        - services
        - createdAt
        - updatedAt

    UpdateDivisionBody:
      type: object
      properties:
        name:
          type: string
        tagline:
          type: ["string", "null"]
        description:
          type: string
        bannerColor:
          type: string
        comingSoon:
          type: boolean
        imageUrl:
          type: ["string", "null"]

    Homepage:
      type: object
      properties:
        id:
          type: integer
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        motto:
          type: string
        missionStatement:
          type: string
        visionStatement:
          type: string
        coreValues:
          type: array
          items:
            type: string
        heroImageUrl:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        googleAnalyticsId:
          type: ["string", "null"]
        updatedAt:
          type: string
      required:
        - id
        - heroTitle
        - heroSubtitle
        - motto
        - missionStatement
        - visionStatement
        - coreValues
        - updatedAt

    UpdateHomepageBody:
      type: object
      properties:
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        motto:
          type: string
        missionStatement:
          type: string
        visionStatement:
          type: string
        coreValues:
          type: array
          items:
            type: string
        heroImageUrl:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        googleAnalyticsId:
          type: ["string", "null"]

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string

    GalleryItem:
      type: object
      required: [id, imageUrl, sortOrder, createdAt]
      properties:
        id:
          type: integer
        divisionSlug:
          type: ["string", "null"]
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string

    CreateGalleryItemBody:
      type: object
      required: [imageUrl]
      properties:
        divisionSlug:
          type: ["string", "null"]
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: integer

    Product:
      type: object
      required: [id, divisionSlug, name, sortOrder, createdAt]
      properties:
        id:
          type: integer
        divisionSlug:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string

    CreateProductBody:
      type: object
      required: [divisionSlug, name]
      properties:
        divisionSlug:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        sortOrder:
          type: integer

    Testimonial:
      type: object
      required: [id, authorName, content, sortOrder, createdAt]
      properties:
        id:
          type: integer
        authorName:
          type: string
        company:
          type: ["string", "null"]
        content:
          type: string
        avatarUrl:
          type: ["string", "null"]
        divisionSlug:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string

    CreateTestimonialBody:
      type: object
      required: [authorName, content]
      properties:
        authorName:
          type: string
        company:
          type: ["string", "null"]
        content:
          type: string
        avatarUrl:
          type: ["string", "null"]
        divisionSlug:
          type: ["string", "null"]
        sortOrder:
          type: integer

    NewsletterSubscribeBody:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    NewsletterSubscriber:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string

    Message:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - message
        - read
        - createdAt

    CreateMessageBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - message

    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        username:
          type: string
      required:
        - success
        - username

    AdminSession:
      type: object
      properties:
        loggedIn:
          type: boolean
        username:
          type: ["string", "null"]
      required:
        - loggedIn

    AdminStats:
      type: object
      properties:
        totalDivisions:
          type: integer
        totalMessages:
          type: integer
        unreadMessages:
          type: integer
        totalServices:
          type: integer
      required:
        - totalDivisions
        - totalMessages
        - unreadMessages
        - totalServices