# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fredora Multiconcept API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  divisionSlug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  price?: string | null;
  sortOrder?: number;
}