# lib/api-zod/src/generated/types/createTestimonialBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fredora Multiconcept API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTestimonialBody {
  authorName: string;
  /** @nullable */
  company?: string | null;
  content: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  divisionSlug?: string | null;
  sortOrder?: number;
}