# lib/api-zod/src/generated/types/newsletterSubscribeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fredora Multiconcept API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterSubscribeBody {
  email: string;
}